# src/openapi/dist/openapi.yaml
openapi: 3.1.0
info:
  title: mucajey Card API
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.mucajey.twicemind.com
    description: Production server
paths:
  /{filename}/spotify-sync:
    post:
      summary: Spotify Playlist Sync
      description: Spotify Playlist Sync
      tags:
        - Sync
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
  /{filename}/itunes-sync:
    post:
      summary: iTunes Sync
      description: iTunes Sync
      tags:
        - Sync
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
...
  /{edition}/{cardId}/apple:
    get:
      summary: Get only Apple ID
      description: Get only Apple ID
      tags:
        - Scan
      parameters:
        - name: edition
          in: path
          required: true
          schema:
            type: string
        - name: cardId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
...
  /year/{year}:
    get:
      summary: Filter cards by year
      description: Filter cards by year
      tags:
        - Cards
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
        '500':
          description: Internal Server Error
...
  /card/year/{year}:
    get:
      summary: List all cards from the specified year.
      description: List all cards from the specified year.
      tags:
        - Card
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
...
  /card/{edition}/{id}:
    patch:
      summary: Update a card by edition and id.
      description: Update a card by edition and id.
      tags:
        - Card
      parameters:
        - name: edition
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
        '404':
          description: Not Found
    delete:
      summary: Delete a card by edition and id.
      description: Delete a card by edition and id.
      tags:
        - Card
      parameters:
        - name: edition
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
        '404':
          description: Not Found
...